requirements.txt
Django==4.1
djangorestframework==3.14
djangorestframework-simplejwt==5.2.1
django-filter==2.4.0


